requirements.txt
flask==2.1.0
boto3==1.17.52
pyJWT==1.7.1